# Optimise parallelism in compute cells operations by zipping first