--- a/beacon_node/beacon_chain/src/kzg_utils.rs
+++ b/beacon_node/beacon_chain/src/kzg_utils.rs
@@ -187,7 +187,7 @@ pub fn blobs_to_data_column_sidecars<E: EthSpec>(
         .collect::<Vec<_>>();
 
     // NOTE: assumes blob sidecars are ordered by index
-    let zipped: Vec<_> = blobs.into_iter().zip(proof_chunks.into_iter()).collect();
+    let zipped: Vec<_> = blobs.iter().zip(proof_chunks).collect();
     let blob_cells_and_proofs_vec = zipped
         .into_par_iter()
         .map(|(blob, proofs)| {